requirements.txt
streamlit
requests
fastapi
python-dotenv
pillow
pymongo
celery
redis
PyMuPDF